# internal/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Subscription Aggregation API
  description: |
    REST-сервис для агрегации данных об онлайн-подписках пользователей.
    Тестовое задание Junior Golang Developer (Effective Mobile).
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
security:
  - BearerAuth: []
paths:
...
  /users/{id}:
    get:
      tags: [Users]
      summary: Получить пользователя
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    put:
      tags: [Users]
      summary: Обновить пользователя
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        "200":
          description: Обновлено

    delete:
      tags: [Users]
      summary: Удалить пользователя
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        "204":
          description: Удалено

  # ---------------- SUBSCRIPTIONS ----------------
  /subscriptions:
    post:
      tags: [Subscriptions]
      summary: Создать подписку
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubscriptionRequest'
      responses:
        "201":
          description: Подписка создана

    get:
      tags: [Subscriptions]
      summary: Получить список подписок
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
  /subscriptions/total:
    get:
      tags: [Subscriptions]
      summary: Подсчёт суммарной стоимости подписок
      parameters:
        - name: user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: service_name
          in: query
          schema:
            type: string
        - name: from
          in: query
          example: "01-2025"
        - name: to
          in: query
          example: "12-2025"
      responses:
        "200":
          description: Итоговая сумма
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TotalResponse'
  /subscriptions/{id}:
    get:
      tags: [Subscriptions]
      summary: Получить подписку
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'

    put:
      tags: [Subscriptions]
      summary: Обновить подписку (admin)
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        "200":
          description: Обновлено

    delete:
      tags: [Subscriptions]
      summary: Удалить подписку (admin)
      parameters:
        - $ref: '#/components/parameters/ID'
      responses:
        "204":
          description: Удалено

  # ---------------- SERVICES ----------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
    Subscription:
      type: object
      properties:
        id:
          type: string
          format: uuid
        service_name:
          type: string
        price:
          type: integer
          example: 400
        user_id:
          type: string
          format: uuid
        start_date:
          type: string
          example: "07-2025"
        end_date:
          type: string
          nullable: true
    CreateSubscriptionRequest:
      type: object
      required: [service_name, price, user_id, start_date]
      properties:
        service_name:
          type: string
        price:
          type: integer
        user_id:
          type: string
          format: uuid
        start_date:
          type: string
          example: "07-2025"
    TotalResponse:
      type: object
      properties:
        total:
          type: integer
          example: 1200
...